models/photo-organizer/requirements.txt
streamlit
facenet_pytorch
fastai>=2.5
pillow